[Router] Move Anthropic stream diagnostics to their own file
